refactor(api): split facts DTO ownership
Move facts-owned DTOs into focused API modules while keeping the monolithic API types module as a compatibility re-export.

Tasks: T005, T006, T007
