+++ cardano-mpfs-api/lib/Cardano/MPFS/API/Types.hs
@@ -85,12 +85,9 @@ import Data.Aeson
     , (.:)
     , (.=)
     )
-import Data.ByteString (ByteString)
-import Data.ByteString.Base16 qualified as B16
 import Data.Proxy (Proxy (..))
 import Data.Swagger
-    ( ToParamSchema (..)
-    , ToSchema (..)
+    ( ToSchema (..)
     , declareSchemaRef
     , description
     , properties
@@ -99,13 +96,21 @@ import Data.Swagger
 import Data.Swagger qualified as Swagger
 import Data.Swagger.Declare (Declare)
 import Data.Text (Text)
-import Data.Text qualified as T
-import Data.Text.Encoding qualified as TE
 import Data.Word (Word64)
 import GHC.IsList (IsList (..))
-import Servant.API (FromHttpApiData (..))
 
 import Cardano.MPFS.API.Encoding (Hex (..))
+import Cardano.MPFS.API.Types.Common
+    ( ChainPointJSON (..)
+    , TokenIdJSON (..)
+    , UnverifiedPParams (..)
+    , UtxoEntry (..)
+    , UtxoEntryRefOnly (..)
+    , UtxoRef (..)
+    , UtxoSetWitness (..)
+    , VerificationSnapshot (..)
+    )
+import Cardano.MPFS.API.Types.Facts (BootFacts (..))
 
 -- | Response for @GET \/status@.
 data StatusResponse = StatusResponse
@@ -145,79 +150,6 @@ instance FromJSON StatusResponse where
             <*> o .: "checkpoint_slot"
             <*> o .: "checkpoint_block_id"
             <*> o .: "utxo_root"
-
--- | Indexed chain point baked into proof-bearing
--- responses. Identifies the exact block at which
--- the bundled proofs are valid.
-data ChainPointJSON = ChainPointJSON
-    { cpSlot :: Word64
-    -- ^ Slot of the indexed chain point
-    , cpBlockId :: Hex
-    -- ^ Block hash at that slot (hex)
-    }
-    deriving (Eq, Show)
-
-instance ToJSON ChainPointJSON where
-    toJSON ChainPointJSON{..} =
-        object
-            [ "slot" .= cpSlot
-            , "block_id" .= cpBlockId
-            ]
-
-instance FromJSON ChainPointJSON where
-    parseJSON = withObject "ChainPointJSON" $ \o ->
-        ChainPointJSON
-            <$> o .: "slot"
-            <*> o .: "block_id"
-
--- | Verification snapshot: the exact UTxO-CSMT root
--- plus indexed chain point against which a
--- proof-bearing response's bundled proofs are valid.
--- Every proof-bearing response embeds one of these so
--- clients can verify offline without a separate
--- @GET \/status@ call.
-data VerificationSnapshot = VerificationSnapshot
-    { vsUtxoRoot :: Hex
-    -- ^ UTxO-CSMT root hash
-    , vsChainPoint :: ChainPointJSON
-    -- ^ Indexed chain point
-    }
-    deriving (Eq, Show)
-
-instance ToJSON VerificationSnapshot where
-    toJSON VerificationSnapshot{..} =
-        object
-            [ "utxo_root" .= vsUtxoRoot
-            , "chainpoint" .= vsChainPoint
-            ]
-
-instance FromJSON VerificationSnapshot where
-    parseJSON =
-        withObject "VerificationSnapshot" $ \o ->
-            VerificationSnapshot
-                <$> o .: "utxo_root"
-                <*> o .: "chainpoint"
-
--- | Hex-encoded token identifier for JSON transport.
-newtype TokenIdJSON = TokenIdJSON
-    { unTokenIdJSON :: ByteString
-    }
-    deriving (Eq, Show)
-
-instance ToJSON TokenIdJSON where
-    toJSON (TokenIdJSON bs) =
-        toJSON (Hex bs)
-
-instance FromJSON TokenIdJSON where
-    parseJSON value = do
-        Hex bs <- parseJSON value
-        pure (TokenIdJSON bs)
-
-instance FromHttpApiData TokenIdJSON where
-    parseUrlPiece t =
-        case B16.decode (TE.encodeUtf8 t) of
-            Right bs -> Right (TokenIdJSON bs)
-            Left err -> Left (T.pack err)
 
 -- | JSON representation of on-chain token state.
 data TokenStateJSON = TokenStateJSON
@@ -371,112 +303,6 @@ instance FromJSON WitnessedUtxo where
 -- Old types stay until every endpoint is migrated; they
 -- will be removed in the polish slice of #243.
 
--- | UTxO reference (@{ tx_id, tx_ix }@).
-data UtxoRef = UtxoRef
-    { urTxId :: Hex
-    -- ^ Transaction id (32-byte blake2b, hex)
-    , urTxIx :: Word64
-    -- ^ Output index within the transaction
-    }
-    deriving (Eq, Show)
-
-instance ToJSON UtxoRef where
-    toJSON UtxoRef{..} =
-        object
-            [ "tx_id" .= urTxId
-            , "tx_ix" .= urTxIx
-            ]
-
-instance FromJSON UtxoRef where
-    parseJSON = withObject "UtxoRef" $ \o ->
-        UtxoRef
-            <$> o .: "tx_id"
-            <*> o .: "tx_ix"
-
--- | A single attested UTxO: reference, resolved
--- @TxOut@, and a CSMT inclusion proof against the
--- enclosing response's @snapshot.utxo_root@.
-data UtxoEntry = UtxoEntry
-    { ueRef :: UtxoRef
-    -- ^ UTxO reference (@ref@)
-    , ueTxOutCbor :: Hex
-    -- ^ CBOR-encoded @TxOut@ (@txout_cbor@)
-    , ueInclusionProof :: Hex
-    -- ^ CSMT inclusion proof (@inclusion_proof@)
-    }
-    deriving (Eq, Show)
-
-instance ToJSON UtxoEntry where
-    toJSON UtxoEntry{..} =
-        object
-            [ "ref" .= ueRef
-            , "txout_cbor" .= ueTxOutCbor
-            , "inclusion_proof" .= ueInclusionProof
-            ]
-
-instance FromJSON UtxoEntry where
-    parseJSON = withObject "UtxoEntry" $ \o ->
-        UtxoEntry
-            <$> o .: "ref"
-            <*> o .: "txout_cbor"
-            <*> o .: "inclusion_proof"
-
--- | A UTxO entry without a per-entry inclusion proof.
--- Appears inside a 'UtxoSetWitness' where the
--- enclosing 'uswCompletenessProof' attests the whole
--- set at once.
-data UtxoEntryRefOnly = UtxoEntryRefOnly
-    { uerRef :: UtxoRef
-    , uerTxOutCbor :: Hex
-    }
-    deriving (Eq, Show)
-
-instance ToJSON UtxoEntryRefOnly where
-    toJSON UtxoEntryRefOnly{..} =
-        object
-            [ "ref" .= uerRef
-            , "txout_cbor" .= uerTxOutCbor
-            ]
-
-instance FromJSON UtxoEntryRefOnly where
-    parseJSON =
-        withObject "UtxoEntryRefOnly" $ \o ->
-            UtxoEntryRefOnly
-                <$> o .: "ref"
-                <*> o .: "txout_cbor"
-
--- | An enumerated set of UTxOs at a known script-hash
--- prefix together with a single CSMT prefix-completeness
--- proof attesting the set is exactly the leaves under
--- that prefix in the snapshot's CSMT.
---
--- The script-hash prefix itself is not part of the
--- witness — the verifier derives it locally from the
--- trusted blueprint.
-data UtxoSetWitness = UtxoSetWitness
-    { uswEntries :: [UtxoEntryRefOnly]
-    -- ^ Enumerated entries (@entries@)
-    , uswCompletenessProof :: Hex
-    -- ^ CSMT prefix-completeness proof
-    -- (@completeness_proof@)
-    }
-    deriving (Eq, Show)
-
-instance ToJSON UtxoSetWitness where
-    toJSON UtxoSetWitness{..} =
-        object
-            [ "entries" .= uswEntries
-            , "completeness_proof"
-                .= uswCompletenessProof
-            ]
-
-instance FromJSON UtxoSetWitness where
-    parseJSON =
-        withObject "UtxoSetWitness" $ \o ->
-            UtxoSetWitness
-                <$> o .: "entries"
-                <*> o .: "completeness_proof"
-
 -- | Uniform response envelope for proof-bearing write
 -- endpoints under #243.
 --
@@ -525,67 +351,6 @@ instance FromJSON UnsignedTxResponse where
                 <$> o .: "unsigned_tx_cbor"
                 <*> o .: "snapshot"
                 <*> o .: "inputs"
-
--- | Protocol parameters carried by a facts-only boot response.
---
--- The offchain service can report these bytes to downstream
--- tooling, but this slice does not verify their ledger-level
--- meaning. The @verified@ flag is therefore part of the wire
--- contract and must be @false@ for the dummy/test values used
--- by the boot facts verifier.
-data UnverifiedPParams = UnverifiedPParams
-    { uppVerified :: Bool
-    -- ^ Whether the protocol parameters have been verified.
-    , uppCbor :: Hex
-    -- ^ CBOR-encoded protocol parameters (@cbor@).
-    }
-    deriving (Eq, Show)
-
-instance ToJSON UnverifiedPParams where
-    toJSON UnverifiedPParams{..} =
-        object
-            [ "verified" .= uppVerified
-            , "cbor" .= uppCbor
-            ]
-
-instance FromJSON UnverifiedPParams where
-    parseJSON =
-        withObject "UnverifiedPParams" $ \o ->
-            UnverifiedPParams
-                <$> o .: "verified"
-                <*> o .: "cbor"
-
--- | Facts-only boot response.
---
--- Unlike 'UnsignedTxResponse', this envelope carries no
--- transaction CBOR. Its verifier proves only that the
--- advertised wallet UTxOs are included in the trusted
--- snapshot root.
-data BootFacts = BootFacts
-    { bfSnapshot :: VerificationSnapshot
-    -- ^ Snapshot the bundled inclusion proofs target.
-    , bfWalletUtxos :: [UtxoEntry]
-    -- ^ Wallet UTxOs with CSMT inclusion proofs.
-    , bfProtocolParameters :: UnverifiedPParams
-    -- ^ Unverified protocol parameter bytes.
-    }
-    deriving (Eq, Show)
-
-instance ToJSON BootFacts where
-    toJSON BootFacts{..} =
-        object
-            [ "snapshot" .= bfSnapshot
-            , "wallet_utxos" .= bfWalletUtxos
-            , "protocol_parameters" .= bfProtocolParameters
-            ]
-
-instance FromJSON BootFacts where
-    parseJSON =
-        withObject "BootFacts" $ \o ->
-            BootFacts
-                <$> o .: "snapshot"
-                <*> o .: "wallet_utxos"
-                <*> o .: "protocol_parameters"
 
 -- ---------------------------------------------------------
 -- Proof-bearing read responses
@@ -1367,65 +1132,6 @@ instance ToSchema StatusResponse where
                 ?~ "Indexer chain tip, checkpoint, \
                    \and current UTxO-CSMT root"
 
-instance ToSchema ChainPointJSON where
-    declareNamedSchema _ = do
-        word64Schema <-
-            declareSchemaRef (Proxy @Word64)
-        hexSchema <-
-            declareSchemaRef (Proxy @Hex)
-        pure
-            $ Swagger.NamedSchema
-                (Just "ChainPointJSON")
-            $ mempty
-            & Swagger.type_
-                ?~ Swagger.SwaggerObject
-            & properties
-                .~ fromList
-                    [ ("slot", word64Schema)
-                    , ("block_id", hexSchema)
-                    ]
-            & required .~ ["slot", "block_id"]
-            & description
-                ?~ "Indexed chain point baked into \
-                   \proof-bearing responses"
-
-instance ToSchema VerificationSnapshot where
-    declareNamedSchema _ = do
-        hexSchema <-
-            declareSchemaRef (Proxy @Hex)
-        chainPointSchema <-
-            declareSchemaRef (Proxy @ChainPointJSON)
-        pure
-            $ Swagger.NamedSchema
-                (Just "VerificationSnapshot")
-            $ mempty
-            & Swagger.type_
-                ?~ Swagger.SwaggerObject
-            & properties
-                .~ fromList
-                    [ ("utxo_root", hexSchema)
-                    , ("chainpoint", chainPointSchema)
-                    ]
-            & required
-                .~ ["utxo_root", "chainpoint"]
-            & description
-                ?~ "UTxO-CSMT root and indexed chain \
-                   \point against which bundled \
-                   \proofs are valid"
-
-instance ToSchema TokenIdJSON where
-    declareNamedSchema _ =
-        pure
-            $ Swagger.NamedSchema
-                (Just "TokenIdJSON")
-            $ Swagger.toSchema (Proxy @String)
-            & description
-                ?~ "Hex-encoded token identifier"
-
-instance ToParamSchema TokenIdJSON where
-    toParamSchema _ =
-        Swagger.toParamSchema (Proxy @String)
-
 instance ToSchema TokenStateJSON where
     declareNamedSchema _ = do
         textSchema <-
@@ -1820,106 +1526,6 @@ instance ToSchema WitnessedRequest where
             & description
                 ?~ "Pending request plus UTxO witness"
 
--- Post-split (#243) primitives.
-
-instance ToSchema UtxoRef where
-    declareNamedSchema _ = do
-        hexSchema <-
-            declareSchemaRef (Proxy @Hex)
-        word64Schema <-
-            declareSchemaRef (Proxy @Word64)
-        pure
-            $ Swagger.NamedSchema (Just "UtxoRef")
-            $ mempty
-            & Swagger.type_
-                ?~ Swagger.SwaggerObject
-            & properties
-                .~ fromList
-                    [ ("tx_id", hexSchema)
-                    , ("tx_ix", word64Schema)
-                    ]
-            & required .~ ["tx_id", "tx_ix"]
-            & description ?~ "UTxO reference"
-
-instance ToSchema UtxoEntry where
-    declareNamedSchema _ = do
-        refSchema <-
-            declareSchemaRef (Proxy @UtxoRef)
-        hexSchema <-
-            declareSchemaRef (Proxy @Hex)
-        pure
-            $ Swagger.NamedSchema (Just "UtxoEntry")
-            $ mempty
-            & Swagger.type_
-                ?~ Swagger.SwaggerObject
-            & properties
-                .~ fromList
-                    [ ("ref", refSchema)
-                    , ("txout_cbor", hexSchema)
-                    , ("inclusion_proof", hexSchema)
-                    ]
-            & required
-                .~ [ "ref"
-                   , "txout_cbor"
-                   , "inclusion_proof"
-                   ]
-            & description
-                ?~ "UTxO ref, CBOR body, and CSMT \
-                   \inclusion proof against the \
-                   \enclosing snapshot's utxo_root"
-
-instance ToSchema UtxoEntryRefOnly where
-    declareNamedSchema _ = do
-        refSchema <-
-            declareSchemaRef (Proxy @UtxoRef)
-        hexSchema <-
-            declareSchemaRef (Proxy @Hex)
-        pure
-            $ Swagger.NamedSchema
-                (Just "UtxoEntryRefOnly")
-            $ mempty
-            & Swagger.type_
-                ?~ Swagger.SwaggerObject
-            & properties
-                .~ fromList
-                    [ ("ref", refSchema)
-                    , ("txout_cbor", hexSchema)
-                    ]
-            & required .~ ["ref", "txout_cbor"]
-            & description
-                ?~ "UTxO ref and CBOR body, without a \
-                   \per-entry inclusion proof. Used \
-                   \inside a UtxoSetWitness."
-
-instance ToSchema UtxoSetWitness where
-    declareNamedSchema _ = do
-        entrySchema <-
-            declareSchemaRef
-                (Proxy @[UtxoEntryRefOnly])
-        hexSchema <-
-            declareSchemaRef (Proxy @Hex)
-        pure
-            $ Swagger.NamedSchema
-                (Just "UtxoSetWitness")
-            $ mempty
-            & Swagger.type_
-                ?~ Swagger.SwaggerObject
-            & properties
-                .~ fromList
-                    [ ("entries", entrySchema)
-                    , ("completeness_proof", hexSchema)
-                    ]
-            & required
-                .~ [ "entries"
-                   , "completeness_proof"
-                   ]
-            & description
-                ?~ "Enumerated UTxOs at a known \
-                   \script-hash prefix, with a single \
-                   \CSMT prefix-completeness proof \
-                   \attesting the set is exactly the \
-                   \leaves under that prefix."
-
 instance ToSchema UnsignedTxResponse where
     declareNamedSchema _ = do
         hexSchema <-
@@ -1953,59 +1559,6 @@ instance ToSchema UnsignedTxResponse where
                    \snapshot and a flat list of spent \
                    \and reference inputs, each with its \
                    \CSMT inclusion proof."
-
-instance ToSchema UnverifiedPParams where
-    declareNamedSchema _ = do
-        hexSchema <-
-            declareSchemaRef (Proxy @Hex)
-        boolSchema <-
-            declareSchemaRef (Proxy @Bool)
-        pure
-            $ Swagger.NamedSchema
-                (Just "UnverifiedPParams")
-            $ mempty
-            & Swagger.type_
-                ?~ Swagger.SwaggerObject
-            & properties
-                .~ fromList
-                    [ ("verified", boolSchema)
-                    , ("cbor", hexSchema)
-                    ]
-            & required .~ ["verified", "cbor"]
-            & description
-                ?~ "CBOR-encoded protocol parameters \
-                   \reported as unverified facts."
-
-instance ToSchema BootFacts where
-    declareNamedSchema _ = do
-        snapshotSchema <-
-            declareSchemaRef
-                (Proxy @VerificationSnapshot)
-        walletSchema <-
-            declareSchemaRef (Proxy @[UtxoEntry])
-        ppSchema <-
-            declareSchemaRef (Proxy @UnverifiedPParams)
-        pure
-            $ Swagger.NamedSchema (Just "BootFacts")
-            $ mempty
-            & Swagger.type_
-                ?~ Swagger.SwaggerObject
-            & properties
-                .~ fromList
-                    [ ("snapshot", snapshotSchema)
-                    , ("wallet_utxos", walletSchema)
-                    , ("protocol_parameters", ppSchema)
-                    ]
-            & required
-                .~ [ "snapshot"
-                   , "wallet_utxos"
-                   , "protocol_parameters"
-                   ]
-            & description
-                ?~ "Facts-only boot response. Carries \
-                   \wallet UTxO witnesses and \
-                   \unverified protocol parameters, \
-                   \with no unsigned transaction CBOR."
 
 instance ToSchema FactWitness where
     declareNamedSchema _ = do
